new RPC implementation seems to be working. Not fully tested"

+++ clipper_server/src/features.rs
@@ -152,7 +152,7 @@ fn feature_worker(name: String,
                   predictions_counter: Arc<metrics::Counter>,
                   batch_size: usize) {
 
-    let input_type = rest::InputType::Float(784);
+    let input_type = rest::InputType::Integer(784);
     // if the batch_size is less than 1 (these are unsigned
     // integers, so that means batch size == 0), we assume dynamic batching
     let dynamic_batching = batch_size < 1;
